python: disable build of readline plugin
for some unresearched reason, after the build of the readline plugin,
the python binary segfaults.
since we don't need the readline plugin, we can just disable it for the
moment. this issue causes build error of python if libedit is installed
prior to its compilation.

fixes #339

if it turns out that readline support is needed for something, we need
to investigate the issue properly and add libedit/readline as a hard
dependency.