gdb76: search debug files in /lib/debug
the pathsearch logic of gdb is broken when used with a global debug
dir if the packages get symlinked from /opt/pkg into /.
theoretically we could put debug files into bin/.debug and lib/.debug,
but this makes removal harder if unneeded and is non-obvious due to
the use of the "hidden" name .debug.

this is the first step needed to support debugging via gnu debuglink
and separate debug files rather than debug info hardcoded into the
executables.
the advantage is that the binaries come stripped by default and consume
less memory, while the debug info is available when gdb is launched.