butch template: split debug info into separate files if building with -g
debug info is now split off binaries and libraries and placed into /lib/debug,
where gdb will search for them if a program is being debugged.
this makes the binaries basically the same size than stripped ones and we don't
waste memory for the debug info if not used actively.
on the gdb side a patch was necessary, which was committed in adb55cc2c.