wpa_connect_action.sh: fork and recall script to avoid blocking
since the script is called from wpa supplicant in a blocking manner,
the interface is usually not yet setup properly at the time the script
is called so it will fail.
avoid this by forking and calling the script another time with a
short delay so wpa_supplicant has time to finish its work.