add nodejs4 (current LTS version) and bump nodejs to latest
the original nodejs recipe is kept, because v4.x and higher require a newer GCC
than 4.7, which gets built by default in sabotage.

I've added an extra check for the GCC version at the very start of the
build. Otherwise the configure script would only print a notice, and the build
fails somehwere in the middle without throwing an error related to the
GCC version.

Built successfully with GCC 5.2.0.

the nodejs version bump fixes a vuln:
CVE-2015-8027 Denial of Service Vulnerability
https://nodejs.org/en/blog/vulnerability/december-2015-security-releases/