add node.js 4 (current LTS version)
I've kept the other nodejs recipe, because v4.x and higher require a newer GCC
than 4.7, which gets built by default in sabotage.

Also I've added an extra check for the GCC version at the very start of the
build. Otherwise the configure script would only print a notice, and the build
fails somehwere in the middle without throwing an error related to the GCC version.

Built successfully with GCC 5.2.0 on my machine.