gcc3: supply gcc-[ar,nm,ranlib] for LTO compat
in order for LTO to work with static archives (*.a), the tools need
to be aware of the linker plugin. in newer gccs that support lto this
is done via the gcc-* wrappers, which call the real binutils tools
with the right command line options to load the lto plugin.
for example gcc-ar invokes ar with
--plugin=/lib/gcc/ARCH/5.2.0/liblto_plugin.so or similar.
in order to be able to use gcc-ar etc in all recipes, the stage0
toolchain provides them as well as simple symlinks to the original tools.