binutils: make it possible to build dynamic linked
if the environment variable BINUTILS_DYNAMIC is set to 1, binutils
will be built dynamically, which is required to use GCC's LTO plugin
which is loaded via dlopen, as musl does not support dlopen in static
builds.
GCC must be built dynamically as well to make it produce the
file liblto_plugin.so.