binutils: create dummy link to liblto_plugin.so
binutils searches for plugins to load in
$(readlink -f $ARGV0)/../lib/bfd-plugins. due to this, if we later
build a gcc that installs its liblto-plugin.so to /lib/bfd-plugins,
binutils will not pick it up since it looks in /opt/binutils/lib.
thus we create a symlink pointing to /lib/bfd-plugins/liblto_plugin.so,
initally being a dead link, which can be replaced later by the toolchains
plugin if gcc is built dynamically.
this way we dont need to copy stuff into binutils' private dir from the gcc
buildscript which would appear very hackish.

the advantage of having binutils detect liblto-plugin.so automatically
is that we don't have to replace all occurences of ar and ranlib calls with
calls to gcc-ar/ranlib. once the plugin link is pointing to the real plugin,
standard ar and ranlib will do the right thing and we dont need to build all
objects with -ffat-lto-objects.