Revert "pkgconf: do not search host path if PKG_CONFIG_PATH is given"
This reverts commit 4d8c93541b323504900b4acd0fb59ba4b0e2d9b3.

The commit was made under the assumption that if the user sets
PKG_CONFIG_PATH, he explicitly wants to search only in that path
(for example as a means to specify the pkgconfig directory of a
crosscompile sysroot).
however the environment variable that's planned for this usage is
PKG_CONFIG_LIBDIR. if set, it overrides the default /lib/pkgconfig,
and it is used *after* searching in PKG_CONFIG_PATH, which takes
precedence.
some build systems use PKG_CONFIG_PATH to point to local copies of
libraries, and those were broken by this change, for example netsurf.