remove support for grsec
since the announcement that grsec stable will no longer be available to
non-sponsors, it's gotten a PITA to keep it up-to-date because even the
test releases are no longer available on a regular basis.
if someone wants to step in and maintain it, we can revive the grsec support
at another time.