kernel: introduce usage of incremental updates via official patches
downloading the entire kernel tarball (70MB+) just for the sake of
a minor update is a major annoyance for people with slow internet
link.
now we take advantage of the incremental patches offered on
kernel.org, which should make kernel updates a much smoother experience.

for every minor update we now need to add a pkg containing
the download information (pkg/kernel-patch-x.y.z-p), add the version
number to KEEP/kernel.patches and the name of the new kernel-patch pkg
to pkg/kernel-patches.