kernel: build wireless regulatory database into the kernel
the latest kernels now require access to a regulatory domain database [0]
to be able to change the regdom (via iw reg set COUNTRYCODE or by putting
COUNTRY=XX into wpa_supplicant.conf).

there are 2 methods to supply this database:

1) a program called crda[1], which is called by udev on demand and serves
   the entries from a *signed* version of the database.
   since it is written in python, and makes use of crypto, this adds a
   heavy dependency chain just to be able to change the regdom:
   python, openssl, python crypto libs, udev (which is now part of systemd)...

2) an unencrypted database that is baked into the kernel.

we chose of course the latter method.
the regulatory database was copied from openwrt [2], and enhanced with a
new countrycode "HX" which enables all channels and much higher txpower.
this enables one to legally use for example channel 14 with 1W txpower when
on a ship in international waters.
it is of course illegal to use this countrycode to circumvent restrictions
in any existing country.

[0] https://wireless.wiki.kernel.org/en/developers/regulatory/wireless-regdb
[1] https://wireless.wiki.kernel.org/en/developers/regulatory/crda
[2] http://git.openwrt.org/?p=openwrt.git;a=blob;f=package/kernel/mac80211/files/regdb.txt