butch: add butch-pack and butch-unpack for binary archives
sometimes building a package may not be practical on a given system.
compilation may require more RAM than available, or it just takes a
huge amount of time.
the user can now build such a package on a fast system, and create
a binary archive of the result via `butch pack`.
after transferring the archive to the target, it can be installed
via `butch unpack`.
optimally the build and install pc have the same versions of installed
packages, in order to prevent issues with ABI changes or version
bumps in dynamic libraries.
to automate installation of several packages at once, shell primitives
like a for loop can be used:
ex: for pkg in *.tar ; do butch unpack $pkg ; done