add butch-ucheck upstream update checker utility
usage:
butch ucheck libevent libxml2 nano firefox
butch ucheck /src/pkg/*

ucheck needs a working directory, typically the directory you're
currently in.
it fetches the "uchkurl" of every package specified (either specified
as such in the package's vars section, or using the path to
the first specified mirror's tarball), rips out a list of hyperlinks
to what appears to be tarballs and saves it in a db.
if an existing list for the package is found, the old list is diffed
against the new list.

the intent is to run the script periodically to see if upstream
added a new tarball to its mirrors.

it is strongly advised to be used in combination with 'tee', since the
output may not fit into the terminals buffer and you may miss important
information.

if a package you care about has a newer version, and you want to update,
do not forget to recompile all users of the package (butch users $pkg),
so regressions can be fixed immediately before issuing a pull request.