gcc474: initial support to bootstrap java toolchain
if the environment variable option_java is set to 1 (requires
prefetching gcc-ecj-45 package if done manually), gcc will be built
including support for java. this consists of gcj, and a minimal
classpath library (libgcj*.so).
this environment can then be used to bootstrap GNU classpath, which
features a more complete java class library, and jamvm as a java vm,
or, for a 100% complete java experience, openjdk7/icedtea.
the latter can be used to bootstrap openjdk8/icedtea.
the ecj dependency is a red-hat-modified version of eclipse's java
compiler. it is usually refered to via ecj_latest.jar, which at the
time of this commit was refering to ecj-45.jar.
the only other jar files on the ftp are ecj-43.jar and ecj-49.jar,
which are older resp. newer versions of ecj; however they cannot
be used to bootstrap GNU classpath as they seem to implement other
incompatible java versions and break its compilation.

the gcc patches were taken from alpine linux, thanks @fabled.
resources used which are of future interest:
http://git.alpinelinux.org/cgit/aports/tree/main/java-gcj-compat/APKBUILD?id=568316df90a20e57f3b5e9ca29b4723c9c223468
http://git.alpinelinux.org/cgit/aports/tree/main/gcc?id=00880c74943bc337f321595ee51652127da58c9f
http://git.alpinelinux.org/cgit/aports/tree/community/openjdk7?id=1301e6ae134309fb1ba5b5167df1ce392f78673f
http://git.alpinelinux.org/cgit/aports/tree/main/java-common/APKBUILD?id=1301e6ae134309fb1ba5b5167df1ce392f78673f

this is the first commit to use the new optional-dependency scheme introduced
in butch@737dcb9e2ce0d3aca9e55a29404df120ea89321c , which is not yet imported
into the butch version used in sabotage as there are likely some improvements
to follow.