add shell implementation of butch-core
this makes it much simpler to implement new features.
for full performance we need jobflow[0], which parallelizes downloads;
but it can be used as well without it, having the effect that downloads
are done sequentially.
jobflow will soon be added as a sabotage package.

butch-core is the main app that will be called from the butch wrapper
as in the past.

butch-core-helper is a background process written in awk that takes
advantage of awk's hashtables to speed up dependency lookups etc.

butch-merge merges a pkg recipe with a butch template.

[0] https://github.com/rofl0r/jobflow