butch-core: fine-tune jobflow params
using -buffered is a bad idea, because it leads to suppression of output
until a job is finished. so if you was about to build a single package with
missing tarball, the line
2016-05-04 19:32:17 downloading foo ...
was only printed when the dl job finished, leaving the impression that
butch was doing nothing.
also turn down spinup delay a bit since it could lead to a noticable delay
before running the first download command.