butch-core-helper: fix compatibility with gawk
gawk, as an optimization, uses buffered output and doesn't flush
automatically when printing to stdout, even in POSIX and traditional
mode. since usage of the gawk-specific keyword fflush doesn't work,
we force the flush via system(""), which luckily is detected by
busybox awk as not needing to start a subshell which would unnecessarily
hurt performance on a sabotage system.
an alternative fix would be to write to stderr and communicate over that
fd.