use new deps.stage0 for stage0 packages
this elegantly solves the problem that sometimes unwillingly stage0 deps
were pulled in, which required to test in the packages build section
if we were not building for stage0 and then doing an early exit.
previously deps.host was also used as a "trick" to achieve similar behavior.

since deps.stage0 is only used for the stage0 builds, those deps are now
completely "invisible" to butch once we're past that stage.