butch-core: prevent buildscripts from accessing stdin
a curious issue happened with "python-pygame":
when it failed to build it would cause butch waiting forever for the
exit of the download process (which is sending "DONE" on completion).
the "DONE" was swallowed by the buildscript, which asked for confirmation
reading input via the raw_input python function.
this is due to our fd 5 used for the communication with the download process
effectively being made stdin (via dup2()) during the
while read line ; do ... done<&5 loop.