butch template: make butch_do_custom_destdir overrideable
additionally, all post-build actions have been disabled, since
knowledge about the files installed by a package is necessary for any of them.
this heavily restricts the functionality offered by sabotage, so if the
installation into a staging dir and the subsequential symlinking step is not
desired, another eventually better approach could look like this:
if butch_do_custom_destdir is set to false, record the choice but use a
destdir anyway, but instead of symlinking it into the root tree, just move
the contents of the destdir into the rootfs after creating filelists, tarballs,
external debug info, etc.