butch: re-implement butch update and add butch outdated commands
the previous update implementation in the butch C version was pretty destructive
since the user had no means of preview of what was being rebuilt and
intervention for problematic cases, or to remove specific packages from the
list.
these cases are for example when packages have conflicting files,
like in the case of busybox and wget, where when both are installed the user
usually prefers /bin/wget pointing to gnu wget.
or packages that were unlinked (possibly ncurses in favor of curses).

thus "butch update" is now an interactive process similar to "git rebase -i"
with the ability to remove things from the rebuild list and re-order items.
"butch outdated" just prints the list of outdated packages.
the old behaviour can be replicated with "butch rebuild `butch outdated`",
however it is not advisable.