butch: early exit for pkgs without build section
the new shell implementation of butch, unlike its predecessor,
did not exit early when encountering a pkg without build section.
it just replaced the build section part of the build template with
an exit 0, however it turned out that this way it wastes time
by unnecessarily unpacking tarball-dependency packages like
"kernel-tarball" or "util-linux-tarball".

closes #452