libtool.cross: don't pull in .so files if -static was passed
for all libs that ship .la files, libtool defaults to pick the
absolute filename of the *shared* library, even if -static was passed.
this of course cannot work (at least in combination with --static,
which prevents libtool from removing the option since it does not know it).
so for example, a libtool link invocation that has -lssl in it, actually
expands it to /lib/libssl.so.
there are 2 simple fixes to the problem, the "softer" one being the one
committed. it just replaces /lib/libssl.so with /lib/libssl.a when -static
was passed.
the other fix, which would be preferable, prohibits changing -lssl into
an absolute filename at all, however this breaks crosscompilation for example
of sdl-mixer, since there's additional yet unidentified logic that tries to
find the physical copy of the lib before adding the -lfoo version to the
command line, and this logic doesn't restrict its search to the passed sysroot.
the effect is that it finds the lib in the hosts' /lib, and adds this as an -L
flag.
the fix would consist of changing the first assignment

hardcode_direct=no

to

hardcode_direct=unsupported

now with the change done, it should be possible to compile *any* program using
libtool statically by putting "-static --static" into LDFLAGS, and simply
calling libtool.cross.gen after configure (the sysroot param can be passed just
like in the autoconf template in KEEP/pkg_skel, and removing the CROSS_COMPILE
test).