cryptsetup: build statically
if encrypted partitions are used, cryptsetup is a critical piece of
infrastructure that's called directly from rc.boot.
depending on the user's filesystem layout, failure of it to execute
(for example due to a libressl soname bump when dynamically linked)
could lead to uncomfortable situations where the user is forced to
repair his system with a pendrive and a barebone busybox environment
(wget to another host in the lan could be an option as well if ssh
doesn't work either for the same reason).
anyway, to prevent the user from the embarassment of an only semi-
functional environment, it is important to have cryptsetup static
linked, even though the resulting 2 binaries are pretty big
(~750 KB each on x86_64).
eventually the space usage could be tuned by using wolfssl in place
of libressl, or by specifying to use the nettle crypto backend
instead.
a user that doesn't need the 'veritysetup' binary, could eventually
remove it as well (untested, potentially called from cryptsetup).