php: bump to 7.0.10 and fix x-compile
php crosscompilation is highly broken.
there are a number of configure test that provide no cache variable
and always fail for crosscompiles.
https://bugs.php.net/bug.php?id=66103

so we have to disable the features that pull in those checks.
further issues are leaking of host include and lib dirs in various sites
due to stupid "find the physical location of a header" configure checks,
hardcoded /usr/include paths etc; and unconditional usage of the just-built
php interpreter.
that means that the x-compile host needs php in its path and a number
of ugly sed fixes to the Makefiles.