libwnck: bump to 2.29.6
this is the last version that doesn't pull in a dependence on GLib's
GObject-Introspection (GIR) stuff, which requires to rebuild all glib-
based packages with introspection enabled and further pulls in a python
dependency (IIRC).
the latest libwnck versions > 2.90 all require GTK+3.

this update fixes a bug in latest lxpanel which causes it to throw out
a lot of glib assertions and then terminating pre-maturely, which causes
it to be restarted in an endless loop. the visible effect is that the panel
flashes on and off, non-sticky desktop items (if spanning for more than a row)
get resorted endlessly because the desktop's visible area changes due to this
repeatedly changing panel, and generally the panel cannot be used, nor programs
launched.