lxde: add lxappearance-obconf
this commit finishes the big LXDE update in the last 14 or so commits.

the update was tackled because there were some annoying bugs in the version
we used earlier, for one the temperature sensor applet of lxpanel was entirely
non-functional, and pcmanfm failed to open directories (opened from the desktop)
on the active desktop.

the only component that has not been updated is lxsession which now has
a hard dependency on dbus.
this is not acceptable so we stick with the version we were already using,
which fortunately is compatible with all the newer components.

we previously held back from updating LXDE, because only a few weeks after
it was first added to sabotage, it was announced on its mailing list that
LXDE would switch to QT.
because almost all gui apps in sabotage are using GTK+2, this means that
a GUI user would now *always* have 2 bloated gui toolkit loaded into memory,
as soon as he launches an app (and QT is even bigger, and way slower to
compile because it's written in C++).
it turned out though that the development of the GTK LXDE components continued
even though the QT version was focused by the main developer.
judging from the download statistics on sourceforge.net, it strongly
appears as if the majority (in the order of 2 magnitudes) of users is still
using the GTK components.