lxde menu-cache rochade
latest libfm depends on menu-cache, and latest menu-cache depends
on libfm (libfm-extra, to be precise, which is installed as part of libfm).
to break out of this circular dependency, we re-introduce the old
menu-cache package, which we build first, followed by libfm, followed by
latest menu-cache.
it will require a clean test build to see if latest menu-cache is ABI-compatible
or not, if not we'll have to rebuild libfm a second time, or drop the new
menu-cache entirely. maybe it is also possible to build only the libfm-extra
part in a first step.