butch-core: implement new pkgver scheme for updating and (un)pack
for the butch update functionality, butch used to compare the hashes
of the installed packages with the hash of the pkgs in the pkg dir.
this has several drawbacks, the major one being that even cosmetic
changes to a pkg caused it to be scheduled for updating.

the new scheme is now that we add a pkgver tag to the vars section.
if the pkgver is missing, it defaults to 1.
this version is now stored into butch.db instead of the package hash.
whenever a package needs a rebuild, for example after a version bump,
the pkgver is increased by one in the package.
this can also be used to schedule a rebuild of users of a library after
a soname bump, like when libressl was updated in 73df8918.
the pkgver is also used by butch pack/unpack.
butch pack now creates a filename $pkg_$arch_$pkgver.tar(.xz, etc).
butch unpack derives the pkgver from the filename and writes it to
butch.db when unpacking.

in order to benefit from this new scheme it is advised that users
do an update before pulling this commit or installing this butch version,
and then switch their butch.db to the new format.

    while read p v ; do printf "%s 1\n" "$p" ; done < /var/lib/butch.db > tmp
    mv tmp /var/lib/butch.db

if they don't, all installed packages will appear outdated when doing
a butch update.
apart from that, everything will continue working as usual.
note that the pkgver is entirely independent from the specific version of a
pkg, for example GCC 4.7.4 vs pkgver=1.