butch-merge: fallback to busybox awk if there's no awk
since gawk overwrites the symlink /bin/awk of busybox, there's a small
window between unlinking it (since it is not in the exception list in the
butch template) to after it's built when there's no awk available.
butch merge and butch relocate are the only components that are called in
this window so they need to be fail-safe.