autoconf package template: remove toolchain name sed hack
the config.sub that's shipped in old source tarballs doesn't know what
musl is and fails with an error if a toolchain with musl triplet is used.
that's why we used the sed hack to make config.sub think that it's building
for a GNU/linux target.
the disadvantage of that hack is that other toolchain components apart
from the compiler are not detected, because there is no x86_64-linux-gnu-ar
in the path, so in some cases this led to usage of host's ar, ranlib, etc
which could fail to deal with the alien architecture produced by the cross-
compiler. so in such cases we used to explicitly override the AR and RANLIB
variables used by make.
the musl-agnostic config.sub is upstreamed since several years and by now
almost all new source packages make use of that.
in the rare case that a new package still ships an old config.sub, the build
will fail very early with an error message from config.sub (it's important
to note from which directory it was invoked!). we can just copy the "good"
config.sub that's stored in KEEP over the old config.sub and fix it like that.