butch-core: do not write tarball-only packages into butch.db
when more than one package uses a specific upstream tarball, our policy
is to make a single "tarball-only" package that consists only of a mirrors
and vars sections.
it's done this way to avoid having multiple packages overwrite each other's
download (because we do downloads in parallel).
an example is "util-linux-tarball", which contains a number of mostly
independent libraries and some tools, where you may want only some or one of the
libraries, but not the tools which are to some degree already available in
busybox.
historically these tarball-packages were also written into butch.db to skip
download/build scripts from running again (this was done unconditionally for
any package).
it however makes more sense to always check the availability of these tarballs,
since packages using it assume the tarball is available.
a possible scenario where this might not be the case is for example when
util-linux-tarball was "installed" in the past and the tarball got deleted or
the pkg was updated to refer to another tarball, so the build of libblkid or
another user would fail.