remove hardcoded "no-unlink" list from build template
the no-unlink list's purpose was to prevent crucial packages from being unlinked
before rebuilding them, because this would render the system unusable (in the
case of busybox/coreutils) or unfit for rebuilding itself (in the case of sed,
awk, gcc, make, etc - these tools are assumed to be always there from numerous
places, even butch itself).

it's more elegant if we can set this special status directly in the affected
package instead of maintaining a separate list.