add initramfs:
This adds an initrd-style pre-userspace to be included in the bootloader
configuration. This is useful when your rootfs requires user-space actions
before being available. Examples:
    - lvm logical volumes (lvm)
    - encrypted partitions (cryptsetup)
    - squashfs rootfs

Step-by-step instruction for usage:
    - Install lvm or cryptsetup (or both), depending on your disk structure
    - Rebuild the initramfs so the executables get included
    - The initramfs is placed as /boot/initrd.img. Add it to your extlinux
        boot configuration with "INITRD initrd.img"
    - Use the kernel options from below to your extlinux config like this:
        "APPEND cryptsetup=/dev/sda2:cryptroot root=/dev/mapper/cryptroot"

The kernel options as defined in bootparam(7) still apply. You are
encouraged to read that manpage first. The options root, rootfstype,
rootfsflags, rw, ro and init are not interpreted by the kernel when an
initramfs is used, so the initramfs is going to use them for
mounting the rootfs.

The following additional kernel options are interpreted by the initramfs:
  boot=<device>
    Mount a block device as temporary /boot. You only need this when
    you want to loop-mount an squashfs as rootfs.

    cryptsetup=<device>:<name>
        Run cryptsetup to decrypt the blockdevice <device> and make it
        available from /dev/mapper/<name>

    lvm
        Before attempting to mount the root, scan for lvm volumes and make
        them available at /dev/mapper/<vol-group>-<vol-name>

    rescue
        Instead of attempting to boot the mounted rootfs, drop an shell

For the options cryptsetup, boot and root, it is also allowed to use syntax
of the format "UUID=<uuid>" or "LABEL=<disk label>". To find those values
for your existing partition, you can use blkid(8)

Example config:

LABEL Sabotage
    KERNEL vmlinuz
    INITRD initrd.img
    APPEND crypt=UUID=a7a....29d:nyu-crypt lvm root=/dev/mapper/vg01-root