xorriso: fix x-compile
-Wl,-static is a trick to work around libtool, which strips off -static.
unfortunately, it breaks cross-compilation with a vanilla musl-cross cross-
compiler, because this option is only passed to the linker, so the compiler
does not know it is compiling a static binary, and thus passes -lgcc_s rather
than -lgcc_eh to the linker. libgcc_s is the dynamic variant of libgcc_eh, so
this will always fail.
our native toolchains have a fix in the gcc spec file that uses libgcc_eh
always, since there are a number of advantages doing so, however vanilla gcc
doesn't, and musl-cross neither.

there's a second trick to workaround libtool insisting on dynamic linking, and
that is --static, which it doesn't recognize either and so doesn't strip it off
the users LDFLAGS. this trick is safe because it is processed by both the
compiler and the linker.