install services from butch template instead manually from each pkg
we now just put service and serviceflags tags into [vars] when
a package ships a service. this allows uniform treatment of
the way the installation is done.

IMPORTANT: with this change, we also remove the default parameter
--force being passed to the install-service script, which had the
nasty effect that recompiling any pkg with a service, for example
wpa_supplicant, would overwrite the eventually customized service
script that was already installed in /etc, and was quite against
our practice of not overwriting things in /etc.

TODO: the "services" pkg still calls install-service for a number
of services. these should be split up into one pkg each and just
have a [vars] section with appropriate service and serviceflags
variables. it also needs to be assured that such a type of pkg
without build section is not skipped.