Revert "cairomm: bump to 1.15.1 [untested]"
This reverts commit b6cba69c53e71fe885781b80dd8da5afa2062eb5.

this verson of cairomm requires libsigc++-3.0 (at least according
to pkg-config, since there is no actual version 3.0+), and the
last sigc++ version, 2.99.7 requires a full C++-14 compiler
(and, judging from the 2.99ish version number,  eventually a gtk+3
base), whereas the latest non-2.99.x version, 2.10.0 requires a full
C++11 compiler. eventually that version's configure test can be
convinced that our GCC 4.7.4 which is 95% C++11 ready, is good enough
and the build will succeed if it doesn't use the most exotic feature.
so before updating cairomm we have to first figure out the last
libsigc++ version that's GCC 4.7-compatible, and then pick the last
cairomm version that fits to that *and* works with all consumers
in the build chain up to inkscape.