autoconf package skeleton: fix configure cross-compile detection
there's a generic issue in autoconf that cross-compiling is not made
entirely dependent on whether --host was given (this only sets the
internal variable on "maybe") but additionally on whether cross-compiled
binaries can be executed. this holds true when crosscompiling i386 binaries
on an amd64 host, that has the dynamic linker for i386 installed as well.