add butch_build_template.txt as a symlink to the real template
the name of the build template is a relic from back when butch was written
and several "demo" templates existed. its name has been a thorn in my eye
ever since, however i refrained from renaming it in order to keep git log
functional. there's however an elegant solution: just add a symlink to the
template with the "proper" name.