build template: globally fix configure x-compile misdetection
generally we follow the view that explicit is better than implicit,
which in this case means the fix should go into every single package using
configure.

however since this affects only a corner-case scenario (see
b86dfc25c31d05723392f6d999247fda2c857386 and
579a37cd38d81c3bc1f7dc342c3248c2df531ce0
for explanations) it seems to be better to fix it globally "hidden" in the
build template, rather than cluttering the 700+ packages with more cross-compile
specific noise, which also would mean a lot of work.

note that this fix still needs to be applied manually to packages that have
several configure scripts.