add gcc 6.3.0
native update from 4.7.4 works flawlessly.
cross-compilation (compiling a native target gcc) requires use of
an existing GCC 6.3.0 cross-compiler (use musl-cross-make for that).

this is the first time that we build a native compiler with a musl
toolchain triplet, our previous approach was to apply the musl patches
to a gnu toolchain triplet. this is not an issue though, as the entire
sabotage framework was already tuned to handle this, due to our cross-
compiling support.