kbd: overhaul and bump to 2.0.4
since kbd's configure grew an idiotic dependency on libpam,
we now just build the loadkeys program on its own with the
host compiler (this also fixes cross-compilation), and then
just run it once to create the keymaps for use with busybox.
circumventing the configure script/Makefiles also has the
advantage that we don't need bison anymore.