xf86-input-evdev: bump to 2.8.4
last version before they switched to libevdev.
IIRC libevdev is targeted at systemd and hard to compile with eudev,
and/or pulls in a lot of new dependencies.
i tried to bootstrap newer versions a while ago but ran into trouble
due to this.
latest xf86-input-evdev which came out today is 2.10.5.
the older versions work perfectly though.