gdb76: fix static build on mips
since the last release, someone added -Wl,--dynamic-list to the build
system for the sake of the entirely GLIBC-specific libthread_db.
this silently broke the static build for some archs, including mips.
this flag has the effect that the compiler thinks static linking happens,
while the linker in fact forces dynamic linking.
due to this we not only get a dynamic binary when we wanted a static one,
binutils doesn't know musl's dynlinker name and so hardcodes
`/usr/lib/libc.so.1' as interpreter name into the binary, which is completely
bogus.