util-linux: fix static build with arbitrary cross-compiler
our native compiler is patched to always use static libgcc_eh.a instead
of libgcc_s.so. unfortunately we don't have that level of control over
crosscompilers, so we cannot outsmart libtool by using --static, which
also outsmarts gcc (it will not be aware that we static link and thus add
libgcc_s to the linker command).
it's better to use our fixed version of libtool that just accepts -static
as it should.

also fix --sbindir.