python: fix remaining x-compile issues
we already had x-compile working; however last i checked there were some
issues with host includes. i adapted the latest patches from openadk (thanks!)
for our use case (we build pgen separately as first step, unlike them).
after that a cross-compiled i386 python worked flawlessly in a chroot.

the only remaining issue was that the modules for zlib, curses, ssl, etc
were not built. it turned out that this is due to a horribly wrong under-
standing of the python guys how compiler toolchains work.

in setup.py, we find this:

    def detect_modules(self):
        # Ensure that /usr/local is always used
        if not cross_compiling:
            add_dir_to_list(self.compiler.library_dirs, '/usr/local/lib')
            add_dir_to_list(self.compiler.include_dirs, '/usr/local/include')
        if cross_compiling:
            self.add_gcc_paths()

add_gcc_paths runs $CC -E -v - </dev/null and parses the returned LIBRARY_PATH
string; which in the case of regular crosscompilers contains only directories
in the crosscompilers own directory.
then they look into these paths and search there for the literal library files.

in our cross-compile setup, we put the rootfs' lib dir into the CC variable like
this:
CC="i486-linux-musl-gcc -L $SYSROOT/lib -isystem $SYSROOT/include"
this is done so we can compile stuff from a rootfs that is not inside the cross-
compiler toolchains path.

later in setup.py comes this:
        for env_var, arg_name, dir_list in (
                ('LDFLAGS', '-R', self.compiler.runtime_library_dirs),
                ('LDFLAGS', '-L', self.compiler.library_dirs),
                ('CPPFLAGS', '-I', self.compiler.include_dirs)):

it looks also into -L directories passed in LDFLAGS.
that is how i fixed the detection of our module libraries in this commit.

so what we can take from this is that the approach used by python to search
for headers and libraries by looking into directories is a completely broken
concept.

the only correct way to test for existence of a library libfoo is:
$CC $CFLAGS $LDFLAGS empty.c -lfoo 2>/dev/null && echo "found" || echo "nope!"

this method for works for native compiles, crosscompiles, macintosh, linux,
AIX, BSD, cygwin, mingw, whatever, regardless of how the user has configured
his toolchain to find its libraries, and regardless of which suffix those files
use, be it .so, .a, .dylib, or even .dll.

likewise, the only valid method for testing if a header foo.h is available is:

echo "#include <foo.h>" > test.c
$CC $CFLAGS $CPPFLAGS $LDFLAGS test.c 2>/dev/null && echo "found" ||echo "nope!"

so regardless how the user has configured his toolchain, be it the compiler's
default paths, usage of -I or -isystem in any of the involved variables, the
compiler will always be able to tell if the header could be located.