gawk: build statically and without libedit
if gawk is installed as the system awk, and libedit is getting rebuilt,
a dynamically linked gawk will cease to work during the rebuild, so
the rebuild, which uses awk as well will fail, so the installed gawk
will be stuck in a broken state. as such a critical piece of infrastructure,
it must be statically linked.