add angr and dependencies
python-z3-angr is probably not needed, i add it anyway for reference.
can be removed later when build was reproduced on another host and confirmed
as unnecessary.

due to some dynloader quirk in the pyvex/unicorn packages, simuvex and thus
angr can only be used when the pyvex python path is in LD_LIBRARY_PATH.
for that reason we supply a wrapper script that lands us either in a REPL,
or can be used to launch an angr script.