firefox: add to no_unlink list
firefox takes such a huge amount of time to build that it can become
very uncomfortable when the build fails and you're left with a system
without browser. you cannot even be sure that you can have your current
version instance running while the rebuild is in process since firefox
tends to crash often due to bugs or due to its huge memory consumption.
fix this situation by adding firefox to the exceptions that are not
removed prior to the build.