wget: bump to 1.19 and upgrade gnulib flak
gnulib authors surpassed themselves in the latest release, so we now
need heavy artillery to keep them from messing with our binaries.
this time they even managed to break things by assuming unlink is
broken, then using a unlink->rpl_unlink macro to point to their own
version, which breaks the struct in ftp.c using "unlink" as the name
of a member.