butch-relocate: reapply the fix for libfm in a safer way
8e37c7dab52736613b783d8d6efe8b313016aab0 reverted the change from
8dea77dad3dceb6ff11218f50c12196cef930a29 , which caused havoc on
one of our systems, it rm -rf'd /opt there.
how that could happen is still not clear, since in a clean world
build the only package hitting that branch is libfm.
that change built on the change from
14dcc0bccee07a4f835e487a106fe8fed87b37ef , which is lacking a detailed
commit message but it appears that it was also related to libfm;
since at that day the whole lxde stack was updated to the latest versions.

it appears therefore as if that change is actually unnecessary since
for the special case in libfm we need to actually just unlink a symlink
to a directory.

in order to be able to find out the cause of the /opt removal on our
system, we now added a guard that will be hit when the target exists
and is a proper directory, and we'll cause the build to fail so we'll
be able to investigate the exact circumstances.
it is very unlikely that any regular user will ever hit this scenario.